Add files via upload
Login.html is a simple web page to show login method. The designing of user interface is done with Bootstrap and made carefully considering user friendly method. It is a simple web page with basic validations which is coded in Angularjs. The webpage contains the fields like user name, email id and password. All the fields have validations which are fall the fields should not be empty, email id should have proper format, and password should have at least 8 characters. Until and unless these conditions are satisfied, Login button will be disabled. Once all the conditions are approved, login is approval.